requirements.txt
# AI Longform YouTube Tool - Requirements
# Python 3.10+

# Web Framework
streamlit>=1.30.0

# AI APIs
anthropic>=0.25.0
httpx>=0.27.0
together>=1.0.0
google-generativeai>=0.3.0

# YouTube API
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.0
isodate>=0.6.0

# TTS
edge-tts>=6.1.0
pydub>=0.25.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0

# Python 3.13 호환성
legacy-cgi>=2.6.0
audioop-lts>=0.2.0
